completed

+++ src/controller/variant.controller.js
@@ -1,22 +1,17 @@
 import { Product } from "../models/product.models.js";
-import { variant } from "../models/variant.models.js";
+import { variant } from "../models/varient.models.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/Apiresponse.js";
-import { asyncHandling } from "../utils/Asynchandler.js";
+import { asynchandling } from "../utils/Asynchandler.js";
 
-const validateVariantData = (name, sku, additional_cost, stock_count) => {
+const addVariant = asynchandling(async (req, res) => {
+  const { productId } = req.params;
+  const { name, sku, additional_cost, stock_count } = req.body;
   if (!(name && sku && additional_cost && stock_count)) {
     throw new ApiError(401, "All fields are required");
   }
-};
-
-const addVariant = asyncHandling(async (req, res) => {
-  const { productId } = req.params;
-  const { name, sku, additional_cost, stock_count } = req.body;
   const product = await Product.findById(productId);
   if (!product) throw new ApiError(400, "Product not found");
-
-  validateVariantData(name, sku, additional_cost, stock_count);
 
   const existingVariant = await variant.findOne({ $or: [{ name }, { sku }] });
   if (existingVariant)
@@ -36,7 +31,7 @@ const addVariant = asyncHandling(async (req, res) => {
     .json(new ApiResponse(200, newVariant, "Variant added successfully"));
 });
 
-const getVariant = asyncHandling(async (req, res) => {
+const getVariant = asynchandling(async (req, res) => {
   const { id } = req.params;
   const variantData = await variant.findById(id);
   if (!variantData) throw new ApiError(400, "Variant not found");
@@ -44,8 +39,21 @@ const getVariant = asyncHandling(async (req, res) => {
   res.status(200).json(new ApiResponse(200, variantData, "Variant found"));
 });
 
-const deleteVariant = asyncHandling(async (req, res) => {
+const deleteVariant = asynchandling(async (req, res) => {
   const { id } = req.params;
+  const { productId } = req.params;
+
+  const product = await Product.findById(productId);
+  if (!product) throw new ApiError(400, "Product not found");
+  const variantIndex = product.variant.indexOf(id);
+  if (variantIndex === -1)
+    throw new ApiError(400, "Variant not found in product");
+
+  if (variantIndex !== -1) {
+    product.variant.splice(variantIndex, 1);
+    await product.save();
+    console.log(product.variant);
+  }
   const deletedVariant = await variant.findByIdAndDelete(id);
   if (!deletedVariant)
     throw new ApiError(400, "Something went wrong while deleting the variant");
@@ -53,15 +61,21 @@ const deleteVariant = asyncHandling(async (req, res) => {
   res.status(200).json(new ApiResponse(200, deletedVariant, "Variant deleted"));
 });
 
-const updateVariant = asyncHandling(async (req, res) => {
-  const { id, name, sku, additional_cost, stock_count } = req.body;
-  validateVariantData(name, sku, additional_cost, stock_count);
-
-  const updatedVariant = await variant.findByIdAndUpdate(
-    id,
-    { name, sku, additional_cost, stock_count },
-    { new: true }
-  );
+const updateVariant = asynchandling(async (req, res) => {
+  const { id } = req.params;
+  const { name, sku, additional_cost, stock_count } = req.body;
+  console.log(req.body);
+  if (!name && !sku && !additional_cost && !stock_count) {
+    throw new ApiError(401, "minimum one field input is required");
+  }
+  let update = {};
+  if (name) update.name = name;
+  if (sku) update.sku = sku;
+  if (additional_cost) update.additional_cost = additional_cost;
+  if (stock_count) update.stock_count = stock_count;
+  const updatedVariant = await variant.findByIdAndUpdate(id, update, {
+    new: true,
+  });
   if (!updatedVariant)
     throw new ApiError(400, "Something went wrong while updating the variant");
